requirements.txt
Django==2.2.9
django-summernote==0.8.11.6
django-taggit==1.2.0
Markdown==3.1.1
Pillow==7.0.0
pytz==2019.3
sqlparse==0.3.0
